requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pandas==2.2.1
scikit-learn==1.4.1.post1
google-genai==0.1.0
python-multipart==0.0.9
pydantic==2.6.4
numpy==1.26.4
gunicorn==21.2.0
